Split implementation of GCS API out from its interface.
